requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
anydex == 0.1.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
auto-python-2014 == 14.1.5

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
auto-python-for-mobile == 14.1.5

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
bayesbridge == 0.2.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
bitmex-ws == 0.4.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
bitmex-ws-onix == 0.3.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
blocked-matrix-utils == 1.0.9

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
bmcs == 0.0.2a46

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
build-py == 1.0.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
buildtools == 1.0.6

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
cgp-dss-data-loader == 1.1.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
clicmod == 0.7.10

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
cmdline-brute == 0.5

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
cninfowebapi == 0.1.4

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
consultor == 0.2.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
cryptools == 0.0.6

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
ddns == 2.9.8

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
dials == 0.0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
django-parakeet == 0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
douyin-downloader == 1.0.4

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
feb == 0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
flask-nsfw == 6

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
flatten-json == 0.1.13

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 100,101
# Project 1/astronomy.ipynb: 100,101
folium == 0.11.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
gardener-cicd-base == 1.1220.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
gcsa == 1.2.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 79,85
# Project 1/astronomy.ipynb: 79,85
geopy == 2.1.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
jw == 1.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 4
# Project 1/astronomy.ipynb: 4
matplotlib == 3.3.3

# Project 1/util/.ipynb_checkpoints/parser-checkpoint.py: 11
# Project 1/util/parser.py: 11
numpy == 1.19.4

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
package-generator == 0.0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 6
# Project 1/astronomy.ipynb: 6
# Project 1/util/.ipynb_checkpoints/parser-checkpoint.py: 5
# Project 1/util/parser.py: 5
pandas == 1.1.5

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
patchman == 1.2.32

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
playground-resumeparser == 0.0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
plda == 1.0.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
prestoadmin == 2.12

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
rapier == 0.0.4

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
sr-python-commons == 0.121

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
stackmob-cli == 0.1.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
stackmob-parse-migrator == 1.0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
streamlink-helper == 0.6

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
swarm-war == 1.0.0b1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
thctools == 1.0.3

# Project 1/util/.ipynb_checkpoints/parser-checkpoint.py: 19
# Project 1/util/parser.py: 19
tqdm == 4.56.0

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
tuido == 0.0.3

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
utillib == 0.2

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
vcv == 3.0.0.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 5
# Project 1/astronomy.ipynb: 5
wordcloud == 1.8.1

# Project 1/.ipynb_checkpoints/astronomy-checkpoint.ipynb: 1
# Project 1/astronomy.ipynb: 1
yang0textutils == 0.0.1
